Create job object on Shell, not Process
